feat: Allow users set upgrade_settings.max_surge property
Azure Kubernetes Service changed the default max surge
in October release, so that if for clusters based
on Kubernetes >1.28 max surge defaults to 10%, see
https://github.com/Azure/AKS/releases/tag/2023-10-01

Previously it was left blank which implied use of
value 1 under the bonnet.

Using the current version of Terraform AzureRM 3.86.0
leads to implicit resetting of the max_surge:

  max_surge = "10%" -> null

The only workaround to avoid such confusing annoyance is to
set the max_surge with explicit value e.g. default "10%".
But, this requires max_surge to be exposed to end-users of
this module. See also
https://github.com/hashicorp/terraform-provider-azurerm/issues/24020

Closes #6

Signed-off-by: Mateusz Łoskot <mateusz@loskot.net>